+++ frontend/src/components/AuthForm/AuthForm.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .auth-form {
   display: flex;
