+++ frontend/src/components/Controler/Controler.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .controler {
   border-bottom-left-radius: 0.5rem;
