+++ frontend/src/components/FormInput/FormInput.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .input-contaier {
   display: flex;
