+++ frontend/src/components/Header/Header.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 header {
   display: flex;
