+++ frontend/src/components/StatusSelector/StatusSelector.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .status-selector {
   border-radius: 0.5rem;
