+++ frontend/src/components/ToDoCard/ToDoCard.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .todo {
   .check {
