+++ frontend/src/routes/Auth/Auth.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@import '../../styles/variables';
 
 .auth-container {
   min-height: 100vh;
