+++ frontend/src/routes/ToDo/ToDo.scss
@@ -1,15 +1,15 @@
-@import './variables';
+@import '../../styles/variables';
 
 .bg-container {
     min-height: 100vh;
-    background-image: url(../assets/bg-mobile-light.jpg);
+    background-image: url(../../assets/bg-mobile-light.jpg);
     background-color: $light-theme-bg;
     background-repeat: no-repeat;
     background-position: top, center;
     padding: 7%;
 
   &.dark-theme{
-    background-image: url(../assets/bg-mobile-dark.jpg);
+    background-image: url(../../assets/bg-mobile-dark.jpg);
     background-color: $dark-theme-bg;
   }
 }
@@ -68,13 +68,13 @@
   }
 
   .bg-container {
-    background-image: url(../assets/bg-desktop-light.jpg);
+    background-image: url(../../assets/bg-desktop-light.jpg);
     display: flex;
     flex-direction: row;
     justify-content: center;
 
     &.dark-theme {
-      background-image: url(../assets/bg-desktop-dark.jpg);
+      background-image: url(../../assets/bg-desktop-dark.jpg);
     }
   }
